Working LoS+spherecasts+delays

+++ Runtime/Scripts/VehicleComponents/Acoustics/ISoundVelocityUser.cs
@@ -4,6 +4,6 @@ namespace Acoustics
 {
     public interface ISoundVelocityUser
     {
-        public void SetSoundVelocity(double vel);     
+        public void SetSoundVelocity(float vel);     
     }
 }
